% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment_CommonFunctions.R
\name{call_RSeQC_modules}
\alias{call_RSeQC_modules}
\title{Call RSeQC modules.}
\usage{
call_RSeQC_modules(bam_per_sample_paths, bed_gene_model_path, output_path,
  read_length, sample_name, wait_flag = NULL)
}
\arguments{
\item{bam_per_sample_paths}{File path for an individual sample .bam file.}

\item{bed_gene_model_path}{Reference gene model in bed format.}

\item{output_path}{Output directory, where RSeQC outputs will be outputted.}

\item{read_length}{Read length.}

\item{sample_name}{Sample name.}

\item{wait_flag}{Call flag with "TRUE", if memory available is <= 50%.}
}
\description{
Call RSeQC modules.
}
