% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment_CommonFunctions.R
\name{call_samtools_sort_index}
\alias{call_samtools_sort_index}
\title{Call samtools sort and index tools.}
\usage{
call_samtools_sort_index(bam_per_sample_paths, output_path, sample_name)
}
\arguments{
\item{bam_per_sample_paths}{File path for an individual sample .bam file.}

\item{output_path}{Output path for sorted and indexed .bam file}

\item{sample_name}{Sample name.}
}
\value{
Runs system commands for sorting and indexing .bam file.
}
\description{
Call samtools sort and index tools.
}
