% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leafcutter_functions.R
\name{create_group_files_multi_pairwisecomp}
\alias{create_group_files_multi_pairwisecomp}
\title{Create LeafCutter group files for multiple pairwise comparisons.}
\usage{
create_group_files_multi_pairwisecomp(df, group_column_name, output_path)
}
\arguments{
\item{df}{Dataframe. Should have a minimum of two columns in the following
order: (i) sample name (filename in the leafcutter
'_perind_numers.count.gz' file) and (ii) grouping variable. Additional
columns will be used as confounders.}

\item{group_column_name}{Character. Name of column with grouping variable.}

\item{output_path}{Character. Output path where _group_file.txt files will be
stored for later LeafCutter differential splicing analyses. Files will be
named using group names for the comparison in question.}
}
\value{
Named _group_file.txt files for LeafCutter differential splicing
  analyses. Files will be named using the groups included in the pairwise
  comparison.
}
\description{
LeafCutter's differential splicing analyses currently only support pairwise
comparisons. For each pairwise comparison it requires a group file to specify
which samples belong to which group. Thus, if a grouping variable contains >
2 groups, multiple pairwise comparisons must be made.
\code{create_group_files_multi_pairwisecomp} will identify the comparisons,
based on an inputted grouping column, and will output separate group .txt
files for each group comparison combination.
}
