% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment_CommonFunctions.R
\name{get_bam_df}
\alias{get_bam_df}
\title{Function to make a data frame of bam paths for sorting and indexing.}
\usage{
get_bam_df(bam_dir_paths)
}
\arguments{
\item{bam_dir_paths}{Path names in which bam files are located.}
}
\value{
Dataframe with file paths.
}
\description{
Function to make a data frame of bam paths for sorting and indexing.
}
