% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment_CommonFunctions.R
\name{get_salmon_parameters_set}
\alias{get_salmon_parameters_set}
\title{Get fixed Salmon parameters.}
\usage{
get_salmon_parameters_set()
}
\value{
String with the fixed salmon parameters.
}
\description{
Function to return a string containing the non-sampleID-dependent salmon
parameters. To be used with salmon quantification.
}
