% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment_CommonFunctions.R
\name{load_sj_df}
\alias{load_sj_df}
\title{Load splice junctions from first pass STAR alignment.}
\usage{
load_sj_df(sj_dir_path)
}
\arguments{
\item{sj_dir_path}{Path to directory containing splice junctions files (SJ.out.tab).}
}
\description{
Load splice junctions from first pass STAR alignment.
}
