% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prealignmentQC_CommonTrimmingFunctions.R
\name{make_results_dir}
\alias{make_results_dir}
\title{Function make directories for results}
\usage{
make_results_dir(results_path, folder_name)
}
\arguments{
\item{results_path}{Character. File path for where output files will be
stored.}

\item{folder_name}{Character. Folder name for output folder.}
}
\value{
Creates a new folder for output (if it doesn't already exist).
}
\description{
Function make directories for results
}
