% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leafcutter_reformat_SJ.R
\name{convert_STAR_SJ_to_junc}
\alias{convert_STAR_SJ_to_junc}
\title{Convert SJ.out.tab files to leafcutter .junc file}
\usage{
convert_STAR_SJ_to_junc(sj_dir_path, output_path)
}
\arguments{
\item{sj_dir_path}{Path to directory containing splice junctions files
(SJ.out.tab).}

\item{output_path}{Path to output directory.}
}
\value{
Leafcutter formatted .junc files in the output_path.
}
\description{
This function will convert SJ.out.tab files outputted by STAR alignment to
the .junc file format necessary for leafcutter splicing analyses.
}
