% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prealignmentQC_CommonTrimmingFunctions.R
\name{get_fastp_df}
\alias{get_fastp_df}
\title{Function to make a data frame of fastQC paths}
\usage{
get_fastp_df(fastq_dir_paths, prefix_to_sample_name = "", fastp_path,
  to_exclude_to_get_sample_name = "_.*")
}
\arguments{
\item{fastq_dir_paths}{Path names in which fastq files for QC are located}

\item{prefix_to_sample_name}{Specify if samples have been given a prefix to
their original name}
}
\value{
Dataframe with file paths.
}
\description{
Function to make a data frame of fastQC paths
}
