% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment_CommonFunctions.R
\name{get_star_parameters_set}
\alias{get_star_parameters_set}
\title{Function to return a string containing the non sampleID dependent STAR parameters set.
To be used in the 1st and 2nd pass alignment.}
\usage{
get_star_parameters_set()
}
\value{
String with the fixed STAR parameters.
}
\description{
Function to return a string containing the non sampleID dependent STAR parameters set.
To be used in the 1st and 2nd pass alignment.
}
